added left-right combiner
if you select two files, one which has the name 'left' and the other 'right', the loader will combine them into one file with the suffix 'cbnd' by linearly interpolating in the gap between the spectra, save to the same directory, and load it. This is an additional if statement in the loader, in the future you could add the functionality of doing multiple left-right pairs.

Prerequisite to using GPU fit on these combined files, as the x axis must be sampled continuously